+++ js/src/main/scala/gpdviz/websocket/WsListener.scala
@@ -3,85 +3,70 @@ package gpdviz.websocket
 import gpdviz.Notif
 import org.scalajs.dom
 import org.scalajs.dom.raw._
+import org.scalajs.dom.window.console
 
 import scala.scalajs.js
+import scala.scalajs.js.timers._
 
-// TODO keep connection alive
 
 class WsListener(handleNotification: (Notif) ⇒ Unit) {
-
-  private val webSocketDiv = dom.document.getElementById("websocket")
 
   private val connectButton = button("connect") { (event: MouseEvent) ⇒
     connect()
     event.preventDefault()
   }
-
-  private val disconnectButton = button("disconnect") { (event: MouseEvent) ⇒
-    disconnect()
-    event.preventDefault()
-  }
-
-  private val messagesDiv = dom.document.createElement("div")
-
   private var wsOpt: Option[WebSocket] = None
+  private var keepAliveHandleOpt: Option[SetIntervalHandle] = None
 
-  webSocketDiv.appendChild(connectButton)
-  webSocketDiv.appendChild(disconnectButton)
-  webSocketDiv.appendChild(messagesDiv)
-
+  dom.document.getElementById("websocket").appendChild(connectButton)
   updateButtons()
+  connect()
 
   private def connect(): Unit = {
-    writeMsg("connecting ...")
+    connectButton.textContent = "Connecting ..."
     connectButton.disabled = true
 
-    val ws = new WebSocket(getWebSocketUri(dom.document))
+    val ws = new WebSocket(getWebSocketUri)
     ws.onopen = { (event: Event) ⇒
-      writeMsg("Connected.")
       wsOpt = Some(ws)
+      keepAliveHandleOpt = Some(setInterval(40 * 1000) {
+        wsOpt foreach { ws ⇒
+          //console.log("ping")
+          ws.send("keep-alive")
+        }
+      })
       updateButtons()
       event
     }
     ws.onerror = { (event: ErrorEvent) ⇒
-      writeMsg(s"Failed: code: ${event.colno}")
-      wsOpt = None
-      updateButtons()
+      console.error(s"Failed: code: ${event.colno}")
+      closed()
     }
-
     ws.onmessage = { (event: MessageEvent) ⇒
       val n = upickle.default.read[Notif](event.data.toString)
       //import pprint.PPrinter.BlackWhite.{apply ⇒ pp}
       //println("onmessage: n = " + pp(n))
       handleNotification(n)
-      //writeMsg(event.data.toString)
     }
-
-    ws.onclose = { (event: Event) ⇒
-      writeMsg("Connection closed")
-      wsOpt = None
-      updateButtons()
+    ws.onclose = { _ ⇒
+      closed()
     }
   }
 
-  private def disconnect(): Unit = {
-    wsOpt foreach(_.close())
+  private def closed(): Unit = {
     wsOpt = None
+    console.warn("Connection closed")
+    keepAliveHandleOpt foreach clearInterval
+    keepAliveHandleOpt = None
     updateButtons()
-    writeMsg("disconnected")
   }
 
   private def updateButtons(): Unit = {
     connectButton.disabled = wsOpt.isDefined
-    disconnectButton.disabled = wsOpt.isEmpty
+    connectButton.textContent = wsOpt.map(_ ⇒ "Connected").getOrElse("Reconnect")
   }
 
-  private def writeMsg(msg: String): Unit = {
-    messagesDiv.innerHTML = msg
-    //org.scalajs.dom.window.console.log(msg)
-  }
-
-  private def getWebSocketUri(document: Document): String = {
+  private def getWebSocketUri: String = {
     val wsProtocol = if (dom.document.location.protocol == "https:") "wss" else "ws"
     s"$wsProtocol://${dom.document.location.host}/ws"
   }