+++ ChangeLog.md
@@ -2,6 +2,9 @@
 
 2017-08-24
 
+- websocket client: connect upon page load and keep connection alive.
+  "Reconnect" button enabled if connection closed by server; this just tries to 
+  re-establish connection. (Complete refresh needs page reload.)
 - refresh initial state of sensor system
 - factor out publish functionality